+++ test.sh
@@ -21,7 +21,8 @@ RUSTFLAGS="-C target-cpu=x86-64 -C target-feature=+sse3" cargo test
 RUSTFLAGS="-C target-cpu=x86-64 -C target-feature=+sse3,+sse4.1" cargo test
 RUSTFLAGS="-C target-cpu=x86-64 -C target-feature=+sse3,+sse4.1,+avx" cargo test
 RUSTFLAGS="-C target-cpu=x86-64 -C target-feature=+sse3,+sse4.1,+avx,+fma" cargo test
-RUSTFLAGS="-C target-cpu=x86-64 -C target-feature=+sse3,+sse4.1,+avx,+fma,+avx2" cargo test
+RUSTFLAGS="-C target-cpu=x86-64 -C target-feature=+sse3,+sse4.1,+avx,+f16c,+fma" cargo test
+RUSTFLAGS="-C target-cpu=x86-64 -C target-feature=+sse3,+sse4.1,+avx,+f16c,+fma,+avx2" cargo test
 RUSTFLAGS="-C target-cpu=x86-64" cargo +nightly test --features="specialization"
 RUSTFLAGS="-C target-cpu=x86-64" cargo +nightly test --features="specialization,no_intrinsics"
 cargo +nightly build --target aarch64-pc-windows-msvc