+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed compilation for ARM targets
 ### Added
 - `LoadFloat3x4`, `StoreFloat3x4`, `LoadFloat4x3`, `StoreFloat4x3`
+- `XMConvertFloatToHalf`, `XMConvertHalfToFloat`
 
 ## [0.1.2] - 2020-06-08
 ### Changed